chore(backend): extract 26 routes into status, runtime, and models routers
Move all inline @app route handlers from create_app() into three
APIRouter modules:

- routes/status.py: GET /, /status, /internal/llama-healthz, /logs (4 routes)
- routes/runtime.py: restart, switch, memory-loading, compatibility,
  power calibration, reset-runtime, cancel-llama (9 routes)
- routes/models.py: register, download, settings, projector, activate,
  move-to-ssd, delete, purge, upload, cancel-upload/download (13 routes)

Use sys.modules["app.main"] reference pattern so monkeypatched
functions resolve at call time — zero test changes needed.

main.py 2328 → 1505 lines.

Refs #67
