+++ app/main.py
@@ -1466,850 +1466,19 @@ def create_app(runtime: RuntimeConfig | None = None, enable_orchestrator: bool |
     if enable_orchestrator is not None:
         app.state.runtime.enable_orchestrator = enable_orchestrator
 
-    @app.get("/", response_class=HTMLResponse)
-    async def root() -> HTMLResponse:
-        return HTMLResponse(_CHAT_HTML_PATH.read_text(encoding="utf-8"))
-
-    @app.get("/status")
-    async def status(request: Request, runtime_cfg: RuntimeConfig = Depends(get_runtime)) -> JSONResponse:
-        download_active, auto_start_remaining = get_status_download_context(request.app, runtime_cfg)
-        return JSONResponse(
-            await build_status(
-                runtime_cfg,
-                app=request.app,
-                download_active=download_active,
-                auto_start_remaining_seconds=auto_start_remaining,
-                system_snapshot=request.app.state.system_metrics_snapshot,
-            )
-        )
-
-    @app.get("/internal/llama-healthz")
-    async def llama_healthz(runtime_cfg: RuntimeConfig = Depends(get_runtime)) -> JSONResponse:
-        transport_healthy = await check_llama_health(runtime_cfg, busy_is_healthy=False)
-        inference_healthy = False
-        if transport_healthy:
-            inference_healthy = await probe_llama_inference_slot(runtime_cfg)
-        return JSONResponse(
-            {
-                "healthy": transport_healthy and inference_healthy,
-                "transport_healthy": transport_healthy,
-                "inference_healthy": inference_healthy,
-            }
-        )
-
-    @app.get("/logs")
-    async def logs() -> StreamingResponse:
-        return StreamingResponse(
-            log_stream(),
-            media_type="text/event-stream",
-            headers={"cache-control": "no-cache", "x-accel-buffering": "no"},
-        )
-
-    @app.post("/internal/restart-llama")
-    async def restart_llama(runtime_cfg: RuntimeConfig = Depends(get_runtime)) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"restarted": False, "reason": "orchestrator_disabled"},
-            )
-        restarted, reason = await restart_managed_llama_process(app)
-        if restarted:
-            return JSONResponse(status_code=200, content={"restarted": True, "reason": reason})
-        return JSONResponse(
-            status_code=200,
-            content={"restarted": False, "reason": reason},
-        )
-
-    @app.post("/internal/llama-runtime/switch")
-    async def switch_llama_runtime(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"switched": False, "reason": "orchestrator_disabled"},
-            )
-
-        payload = await request.json()
-        family = str(payload.get("family") or "").strip()
-        if not family:
-            return JSONResponse(status_code=400, content={"switched": False, "reason": "family_required"})
-
-        slot = find_runtime_slot_by_family(runtime_cfg, family)
-        if slot is None:
-            return JSONResponse(status_code=404, content={"switched": False, "reason": "runtime_not_found"})
-
-        async with app.state.llama_runtime_switch_lock:
-            switch_state = app.state.llama_runtime_switch_state
-            if switch_state.get("active"):
-                return JSONResponse(status_code=409, content={"switched": False, "reason": "switch_already_running"})
-
-            switch_state.update(
-                {
-                    "active": True,
-                    "target_family": family,
-                    "started_at_unix": int(time.time()),
-                    "completed_at_unix": None,
-                    "error": None,
-                }
-            )
-
-            try:
-                restarted, restart_reason = await restart_managed_llama_process(app)
-                install_result = await install_llama_runtime_bundle(runtime_cfg, Path(str(slot["path"])))
-                if not install_result.get("ok"):
-                    reason = str(install_result.get("reason") or "install_failed")
-                    switch_state.update(
-                        {
-                            "active": False,
-                            "completed_at_unix": int(time.time()),
-                            "error": reason,
-                        }
-                    )
-                    return JSONResponse(
-                        status_code=500,
-                        content={
-                            "switched": False,
-                            "reason": reason,
-                            "family": family,
-                            "restarted": restarted,
-                            "restart_reason": restart_reason,
-                        },
-                    )
-
-                marker = write_llama_runtime_bundle_marker(runtime_cfg, slot)
-                switch_state.update(
-                    {
-                        "active": False,
-                        "target_family": None,
-                        "completed_at_unix": int(time.time()),
-                        "error": None,
-                    }
-                )
-                return JSONResponse(
-                    status_code=200,
-                    content={
-                        "switched": True,
-                        "reason": "runtime_switched",
-                        "family": family,
-                        "slot": slot,
-                        "install": install_result,
-                        "restarted": restarted,
-                        "restart_reason": restart_reason,
-                        "marker": marker,
-                    },
-                )
-            except Exception:
-                switch_state.update(
-                    {
-                        "active": False,
-                        "completed_at_unix": int(time.time()),
-                        "error": "switch_failed",
-                    }
-                )
-                raise
-
-    @app.post("/internal/llama-runtime/memory-loading")
-    async def set_llama_memory_loading_mode(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"updated": False, "reason": "orchestrator_disabled"},
-            )
-
-        payload = await request.json()
-        requested_mode = payload.get("mode", payload.get("memory_loading_mode"))
-        mode = normalize_llama_memory_loading_mode(requested_mode)
-        saved = write_llama_runtime_settings(runtime_cfg, memory_loading_mode=mode)
-        restarted, restart_reason = await restart_managed_llama_process(app)
-        memory_loading = build_llama_memory_loading_status(runtime_cfg)
-        return JSONResponse(
-            status_code=200,
-            content={
-                "updated": True,
-                "reason": "memory_loading_updated",
-                "memory_loading": memory_loading,
-                "saved": saved,
-                "restarted": restarted,
-                "restart_reason": restart_reason,
-            },
-        )
-
-    @app.post("/internal/compatibility/large-model-override")
-    async def set_large_model_override(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"updated": False, "reason": "orchestrator_disabled"},
-            )
-
-        payload = await request.json()
-        enabled = normalize_allow_unsupported_large_models(
-            payload.get("enabled", payload.get("allow_unsupported_large_models"))
-        )
-        saved = write_llama_runtime_settings(
-            runtime_cfg,
-            allow_unsupported_large_models=enabled,
-        )
-        compatibility = build_large_model_compatibility(
-            runtime_cfg,
-            model_filename=runtime_cfg.model_path.name,
-            model_size_bytes=(runtime_cfg.model_path.stat().st_size if runtime_cfg.model_path.exists() else None),
-            allow_override=enabled,
-        )
-        return JSONResponse(
-            status_code=200,
-            content={
-                "updated": True,
-                "reason": "large_model_override_updated",
-                "override": {
-                    "enabled": enabled,
-                    "label": (
-                        "Try unsupported large model anyway"
-                        if enabled
-                        else "Use compatibility warnings (default)"
-                    ),
-                },
-                "compatibility": compatibility,
-                "saved": saved,
-            },
-        )
-
-    @app.post("/internal/power-calibration/sample")
-    async def capture_power_calibration_sample(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        payload = await request.json()
-        wall_watts = _safe_positive_float(payload.get("wall_watts"))
-        if wall_watts is None:
-            return JSONResponse(
-                status_code=400,
-                content={"captured": False, "reason": "invalid_wall_watts"},
-            )
-
-        current_power = _build_power_estimate_snapshot()
-        raw_watts = _safe_positive_float(current_power.get("total_watts"))
-        if raw_watts is None:
-            return JSONResponse(
-                status_code=409,
-                content={
-                    "captured": False,
-                    "reason": "power_unavailable",
-                    "power_estimate": build_power_estimate_status(runtime_cfg, current_power),
-                },
-            )
-
-        calibration = _append_power_calibration_sample(
-            runtime_cfg,
-            raw_pmic_watts=raw_watts,
-            wall_watts=wall_watts,
-        )
-        power_status = build_power_estimate_status(runtime_cfg, current_power)
-        return JSONResponse(
-            status_code=200,
-            content={
-                "captured": True,
-                "reason": "power_calibration_sample_captured",
-                "sample": calibration.get("samples", [])[-1] if calibration.get("samples") else None,
-                "calibration": build_power_calibration_status(runtime_cfg),
-                "power_estimate": power_status,
-            },
-        )
-
-    @app.post("/internal/power-calibration/fit")
-    async def fit_power_calibration(
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        ok, reason, calibration = _fit_and_persist_power_calibration(runtime_cfg)
-        status_code = 200 if ok else 400
-        return JSONResponse(
-            status_code=status_code,
-            content={
-                "updated": ok,
-                "reason": reason,
-                "calibration": build_power_calibration_status(runtime_cfg) if ok else calibration,
-            },
-        )
-
-    @app.post("/internal/power-calibration/reset")
-    async def reset_power_calibration(
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        calibration = _reset_power_calibration(runtime_cfg)
-        return JSONResponse(
-            status_code=200,
-            content={
-                "updated": True,
-                "reason": "power_calibration_reset",
-                "calibration": build_power_calibration_status(runtime_cfg),
-                "saved": calibration,
-            },
-        )
-
-    @app.post("/internal/start-model-download")
-    async def start_model_download_now(runtime_cfg: RuntimeConfig = Depends(get_runtime)) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"started": False, "reason": "orchestrator_disabled"},
-            )
-
-        started, reason = await start_model_download(app, runtime_cfg, trigger="manual")
-        status_code = 202 if started else 200
-        return JSONResponse(status_code=status_code, content={"started": started, "reason": reason})
-
-    @app.post("/internal/download-countdown")
-    async def set_download_countdown(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"updated": False, "reason": "orchestrator_disabled"},
-            )
-        payload = await request.json()
-        enabled = bool(payload.get("enabled", True))
-        updated_state = set_download_countdown_enabled(runtime_cfg, enabled)
-        return JSONResponse(
-            status_code=200,
-            content={
-                "updated": True,
-                "reason": "countdown_updated",
-                "countdown_enabled": bool(updated_state.get("countdown_enabled", enabled)),
-            },
-        )
-
-    @app.post("/internal/models/register")
-    async def register_model_endpoint(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        payload = await request.json()
-        source_url = str(payload.get("source_url") or "").strip()
-        alias_raw = payload.get("alias")
-        alias = str(alias_raw).strip() if isinstance(alias_raw, str) and alias_raw.strip() else None
-        ok, reason, model = register_model_url(runtime_cfg, source_url=source_url, alias=alias)
-        if not ok:
-            return JSONResponse(status_code=400, content={"ok": False, "reason": reason})
-        response_payload: dict[str, Any] = {"ok": True, "reason": reason, "model": model}
-        if isinstance(model, dict):
-            model_filename = str(model.get("filename") or "")
-            model_source_url = str(model.get("source_url") or "")
-            size_bytes = 0
-            if model_source_url:
-                size_bytes = await fetch_remote_content_length_bytes(model_source_url)
-            compatibility = build_large_model_compatibility(
-                runtime_cfg,
-                model_filename=model_filename,
-                model_size_bytes=size_bytes or None,
-            )
-            warnings = compatibility.get("warnings")
-            if isinstance(warnings, list) and warnings:
-                response_payload["warnings"] = warnings
-        return JSONResponse(status_code=200, content=response_payload)
-
-    @app.post("/internal/models/download")
-    async def start_selected_model_download(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"started": False, "reason": "orchestrator_disabled"},
-            )
-        payload = await request.json()
-        model_id = str(payload.get("model_id") or "").strip()
-        if not model_id:
-            return JSONResponse(status_code=400, content={"started": False, "reason": "model_id_required"})
-        started, reason = await start_model_download(
-            app,
-            runtime_cfg,
-            trigger="manual-model",
-            model_id=model_id,
-        )
-        status_code = 202 if started else 200
-        return JSONResponse(status_code=status_code, content={"started": started, "reason": reason, "model_id": model_id})
-
-    @app.post("/internal/models/settings")
-    async def update_model_settings_endpoint(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        payload = await request.json()
-        model_id = str(payload.get("model_id") or "").strip()
-        settings = payload.get("settings")
-        if not model_id:
-            return JSONResponse(status_code=400, content={"updated": False, "reason": "model_id_required"})
-        if not isinstance(settings, dict):
-            return JSONResponse(status_code=400, content={"updated": False, "reason": "settings_required"})
-        updated, reason, model = update_model_settings(runtime_cfg, model_id=model_id, settings=settings)
-        if not updated:
-            status_code = 404 if reason == "model_not_found" else 400
-            return JSONResponse(status_code=status_code, content={"updated": False, "reason": reason, "model_id": model_id})
-        restarted = False
-        restart_reason = "not_required"
-        state = ensure_models_state(runtime_cfg)
-        if model_id == str(state.get("active_model_id") or ""):
-            restarted, restart_reason = await restart_managed_llama_process(app)
-        return JSONResponse(
-            status_code=200,
-            content={
-                "updated": True,
-                "reason": reason,
-                "model_id": model_id,
-                "model": model,
-                "restarted": restarted,
-                "restart_reason": restart_reason,
-            },
-        )
-
-    # Settings document routes — extracted to app/routes/settings.py
-
-    @app.post("/internal/models/download-projector")
-    async def download_projector_for_model_endpoint(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        payload = await request.json()
-        model_id = str(payload.get("model_id") or "").strip()
-        if not model_id:
-            return JSONResponse(status_code=400, content={"downloaded": False, "reason": "model_id_required"})
-        downloaded, reason, projector_filename = await asyncio.to_thread(
-            download_default_projector_for_model,
-            runtime=runtime_cfg,
-            model_id=model_id,
-        )
-        if not downloaded:
-            return JSONResponse(
-                status_code=400 if reason != "model_not_found" else 404,
-                content={"downloaded": False, "reason": reason, "model_id": model_id},
-            )
-        state = ensure_models_state(runtime_cfg)
-        model = get_model_by_id(state, model_id)
-        if isinstance(model, dict):
-            settings = normalize_model_settings(model.get("settings"), filename=str(model.get("filename") or ""))
-            settings["vision"]["projector_filename"] = projector_filename
-            model["settings"] = settings
-            save_models_state(runtime_cfg, state)
-        restarted = False
-        restart_reason = "not_required"
-        if str(state.get("active_model_id") or "") == model_id:
-            restarted, restart_reason = await restart_managed_llama_process(app)
-        return JSONResponse(
-            status_code=200,
-            content={
-                "downloaded": True,
-                "reason": reason,
-                "model_id": model_id,
-                "projector_filename": projector_filename,
-                "restarted": restarted,
-                "restart_reason": restart_reason,
-            },
-        )
-
-    @app.post("/internal/models/cancel-download")
-    async def cancel_selected_model_download(runtime_cfg: RuntimeConfig = Depends(get_runtime)) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"cancelled": False, "reason": "orchestrator_disabled"},
-            )
-        cancelled, reason = await cancel_model_download(app, runtime_cfg)
-        return JSONResponse(status_code=200, content={"cancelled": cancelled, "reason": reason})
-
-    @app.post("/internal/models/activate")
-    async def activate_model_endpoint(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"switched": False, "reason": "orchestrator_disabled"},
-            )
-        payload = await request.json()
-        model_id = str(payload.get("model_id") or "").strip()
-        if not model_id:
-            return JSONResponse(status_code=400, content={"switched": False, "reason": "model_id_required"})
-        switched, reason, restarted = await activate_model(app, runtime_cfg, model_id=model_id)
-        status_code = 200 if switched else (409 if reason in {"model_not_ready", "model_not_found"} else 400)
-        return JSONResponse(
-            status_code=status_code,
-            content={"switched": switched, "reason": reason, "restarted": restarted, "model_id": model_id},
-        )
-
-    @app.post("/internal/models/move-to-ssd")
-    async def move_model_to_ssd_endpoint(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"moved": False, "reason": "orchestrator_disabled"},
-            )
-        payload = await request.json()
-        model_id = str(payload.get("model_id") or "").strip()
-        if not model_id:
-            return JSONResponse(status_code=400, content={"moved": False, "reason": "model_id_required"})
-        ssd_dir = get_preferred_model_offload_dir(runtime_cfg)
-        if ssd_dir is None:
-            return JSONResponse(
-                status_code=409,
-                content={"moved": False, "reason": "no_ssd_available", "model_id": model_id},
-            )
-        moved, reason, storage = await asyncio.to_thread(
-            move_model_to_ssd,
-            runtime_cfg,
-            model_id=model_id,
-            ssd_dir=ssd_dir,
-        )
-        if moved:
-            restarted = False
-            restart_reason = "not_required"
-            state = ensure_models_state(runtime_cfg)
-            if model_id == str(state.get("active_model_id") or ""):
-                model = get_model_by_id(state, model_id)
-                if isinstance(model, dict):
-                    runtime_cfg.model_path = resolve_model_runtime_path(runtime_cfg, str(model.get("filename") or ""))
-                restarted, restart_reason = await restart_managed_llama_process(app)
-            return JSONResponse(
-                status_code=200,
-                content={
-                    "moved": True,
-                    "reason": reason,
-                    "model_id": model_id,
-                    "storage": storage,
-                    "restarted": restarted,
-                    "restart_reason": restart_reason,
-                },
-            )
-        status_code = 404 if reason == "model_not_found" else 500 if reason == "move_failed" else 409
-        return JSONResponse(
-            status_code=status_code,
-            content={
-                "moved": False,
-                "reason": reason,
-                "model_id": model_id,
-                "storage": storage,
-                "restarted": False,
-                "restart_reason": "not_required",
-            },
-        )
-
-    @app.post("/internal/models/delete")
-    async def delete_model_endpoint(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"deleted": False, "reason": "orchestrator_disabled"},
-            )
-        payload = await request.json()
-        model_id = str(payload.get("model_id") or "").strip()
-        if not model_id:
-            return JSONResponse(status_code=400, content={"deleted": False, "reason": "model_id_required"})
-        cancelled_download = False
-        async with app.state.download_lock:
-            models_state = ensure_models_state(runtime_cfg)
-            current_download_model_id = str(models_state.get("current_download_model_id") or "").strip()
-            if current_download_model_id == model_id and is_download_task_active(app.state.model_download_task):
-                cancelled_download, cancel_reason = await _cancel_model_download_locked(
-                    app,
-                    runtime_cfg,
-                    expected_model_id=model_id,
-                    timeout_seconds=MODEL_DOWNLOAD_CANCEL_WAIT_TIMEOUT_SECONDS,
-                )
-                if not cancelled_download:
-                    reason = "delete_cancel_timeout" if cancel_reason == "cancel_timeout" else "delete_cancel_failed"
-                    status_code = 409 if reason == "delete_cancel_timeout" else 400
-                    return JSONResponse(
-                        status_code=status_code,
-                        content={
-                            "deleted": False,
-                            "reason": reason,
-                            "model_id": model_id,
-                            "deleted_file": False,
-                            "freed_bytes": 0,
-                            "cancelled_download": False,
-                        },
-                    )
-            deleted, reason, deleted_file, freed_bytes, deleted_active = delete_model(runtime_cfg, model_id=model_id)
-        restarted = False
-        restart_reason = "not_required"
-        if deleted and deleted_active:
-            models_state = ensure_models_state(runtime_cfg)
-            resolve_active_model(models_state, runtime_cfg)
-            restarted, restart_reason = await restart_managed_llama_process(app)
-        if deleted:
-            return JSONResponse(
-                status_code=200,
-                content={
-                    "deleted": True,
-                    "reason": reason,
-                    "model_id": model_id,
-                    "deleted_file": deleted_file,
-                    "freed_bytes": int(max(0, freed_bytes)),
-                    "deleted_active": deleted_active,
-                    "cancelled_download": cancelled_download,
-                    "restarted": restarted,
-                    "restart_reason": restart_reason,
-                },
-            )
-        status_code = 404 if reason == "model_not_found" else 400
-        return JSONResponse(
-            status_code=status_code,
-            content={
-                "deleted": False,
-                "reason": reason,
-                "model_id": model_id,
-                "deleted_file": False,
-                "freed_bytes": 0,
-                "cancelled_download": False,
-            },
-        )
-
-    @app.post("/internal/models/purge")
-    async def purge_models_endpoint(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"purged": False, "reason": "orchestrator_disabled"},
-            )
-        payload = await request.json()
-        reset_bootstrap_flag = bool(payload.get("reset_bootstrap_flag", True))
-        result = await purge_all_models(app, runtime_cfg, reset_bootstrap_flag=reset_bootstrap_flag)
-        return JSONResponse(status_code=200, content=result)
-
-    @app.post("/internal/models/upload")
-    async def upload_model_endpoint(
-        request: Request,
-        runtime_cfg: RuntimeConfig = Depends(get_runtime),
-    ) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"uploaded": False, "reason": "orchestrator_disabled"},
-            )
-
-        raw_filename = request.headers.get("x-potato-filename", "")
-        try:
-            filename = _safe_upload_filename(raw_filename)
-        except ValueError as exc:
-            return JSONResponse(status_code=400, content={"uploaded": False, "reason": str(exc)})
-
-        async with app.state.model_upload_lock:
-            if app.state.model_upload_state.get("active"):
-                return JSONResponse(status_code=409, content={"uploaded": False, "reason": "upload_already_running"})
-
-            declared_total = max(0, _safe_int(request.headers.get("content-length"), 0))
-            max_upload_bytes = get_model_upload_max_bytes()
-            app.state.model_upload_cancel_requested = False
-            app.state.model_upload_state = {
-                "active": True,
-                "model_id": None,
-                "bytes_total": declared_total,
-                "bytes_received": 0,
-                "percent": 0,
-                "error": None,
-            }
-            if max_upload_bytes is not None and declared_total > 0 and declared_total > max_upload_bytes:
-                app.state.model_upload_state.update({"active": False, "error": "upload_too_large"})
-                return JSONResponse(status_code=413, content={"uploaded": False, "reason": "upload_too_large"})
-
-            state = ensure_models_state(runtime_cfg)
-            existing_names = {
-                str(item.get("filename"))
-                for item in state.get("models", [])
-                if isinstance(item, dict)
-            }
-            final_filename = _unique_filename(filename, existing_names)
-            tmp_path = _model_file_path(runtime_cfg, final_filename + ".part")
-            final_path = _model_file_path(runtime_cfg, final_filename)
-            tmp_path.parent.mkdir(parents=True, exist_ok=True)
-            total_received = 0
-            error_reason: str | None = None
-            upload_completed = False
-
-            def _cleanup_partial_upload() -> None:
-                try:
-                    tmp_path.unlink(missing_ok=True)
-                except OSError:
-                    logger.warning("Could not delete partial upload file: %s", tmp_path, exc_info=True)
-
-            try:
-                with tmp_path.open("wb") as handle:
-                    async for chunk in request.stream():
-                        if app.state.model_upload_cancel_requested:
-                            error_reason = "upload_cancelled"
-                            break
-                        if not chunk:
-                            continue
-                        total_received += len(chunk)
-                        if max_upload_bytes is not None and total_received > max_upload_bytes:
-                            error_reason = "upload_too_large"
-                            break
-                        handle.write(chunk)
-                        state_total = app.state.model_upload_state.get("bytes_total") or 0
-                        percent = int((total_received * 100 / state_total)) if state_total else 0
-                        app.state.model_upload_state.update(
-                            {
-                                "bytes_received": total_received,
-                                "percent": max(0, min(100, percent)),
-                            }
-                        )
-
-                if error_reason == "upload_too_large":
-                    _cleanup_partial_upload()
-                    return JSONResponse(status_code=413, content={"uploaded": False, "reason": error_reason})
-                if error_reason == "upload_cancelled":
-                    _cleanup_partial_upload()
-                    return JSONResponse(status_code=200, content={"uploaded": False, "reason": error_reason})
-                if total_received <= 0:
-                    error_reason = "upload_empty"
-                    _cleanup_partial_upload()
-                    return JSONResponse(status_code=400, content={"uploaded": False, "reason": error_reason})
-
-                tmp_path.replace(final_path)
-                models = state.get("models", [])
-                if not isinstance(models, list):
-                    error_reason = "models_state_invalid"
-                    _cleanup_partial_upload()
-                    return JSONResponse(status_code=500, content={"uploaded": False, "reason": error_reason})
-                existing_ids = {
-                    str(item.get("id"))
-                    for item in models
-                    if isinstance(item, dict)
-                }
-                model_id = _unique_model_id(_slugify_id(Path(final_filename).stem), existing_ids)
-                record = {
-                    "id": model_id,
-                    "filename": final_filename,
-                    "source_url": None,
-                    "source_type": "upload",
-                    "status": "ready",
-                    "error": None,
-                }
-                models.append(record)
-                state["active_model_id"] = model_id
-                saved = save_models_state(runtime_cfg, state)
-                runtime_cfg.model_path = final_path
-                restarted, _reason = await restart_managed_llama_process(app)
-                app.state.model_upload_state.update(
-                    {
-                        "active": False,
-                        "error": None,
-                        "model_id": model_id,
-                        "bytes_total": total_received,
-                        "bytes_received": total_received,
-                        "percent": 100,
-                    }
-                )
-                upload_completed = True
-                model = get_model_by_id(saved, model_id)
-                compatibility = build_large_model_compatibility(
-                    runtime_cfg,
-                    model_filename=final_filename,
-                    model_size_bytes=max(total_received, declared_total or 0) or None,
-                )
-                response_payload: dict[str, Any] = {
-                    "uploaded": True,
-                    "model": model,
-                    "switched": True,
-                    "restarted": restarted,
-                }
-                warnings = compatibility.get("warnings")
-                if isinstance(warnings, list) and warnings:
-                    response_payload["warnings"] = warnings
-                return JSONResponse(
-                    status_code=200,
-                    content=response_payload,
-                )
-            except OSError:
-                if error_reason is None:
-                    error_reason = "upload_write_failed"
-                _cleanup_partial_upload()
-                logger.warning("Model upload failed during file write: %s", final_filename, exc_info=True)
-                return JSONResponse(status_code=500, content={"uploaded": False, "reason": error_reason})
-            except Exception:
-                _cleanup_partial_upload()
-                raise
-            finally:
-                if not upload_completed:
-                    app.state.model_upload_state.update(
-                        {
-                            "active": False,
-                            "model_id": None,
-                            "error": error_reason,
-                        }
-                    )
-
-    @app.post("/internal/models/cancel-upload")
-    async def cancel_model_upload(runtime_cfg: RuntimeConfig = Depends(get_runtime)) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"cancelled": False, "reason": "orchestrator_disabled"},
-            )
-        if not app.state.model_upload_state.get("active"):
-            return JSONResponse(status_code=200, content={"cancelled": False, "reason": "not_running"})
-        app.state.model_upload_cancel_requested = True
-        return JSONResponse(status_code=200, content={"cancelled": True, "reason": "cancel_requested"})
-
-    @app.post("/internal/reset-runtime")
-    async def reset_runtime_now(runtime_cfg: RuntimeConfig = Depends(get_runtime)) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"started": False, "reason": "orchestrator_disabled"},
-            )
-
-        started, reason = await start_runtime_reset(runtime_cfg)
-        status_code = 202 if started else 200
-        return JSONResponse(status_code=status_code, content={"started": started, "reason": reason})
-
-    @app.post("/internal/cancel-llama")
-    async def cancel_llama(runtime_cfg: RuntimeConfig = Depends(get_runtime)) -> JSONResponse:
-        if not runtime_cfg.enable_orchestrator:
-            return JSONResponse(
-                status_code=409,
-                content={"cancelled": False, "restarted": False, "reason": "orchestrator_disabled"},
-            )
-
-        cancelled, action = await request_llama_slot_cancel(runtime_cfg)
-        if cancelled:
-            return JSONResponse(
-                status_code=200,
-                content={"cancelled": True, "restarted": False, "method": f"slot:{action}"},
-            )
-
-        return JSONResponse(
-            status_code=200,
-            content={"cancelled": False, "restarted": False, "method": "none", "reason": "slot_action_unavailable"},
-        )
-
-    # Chat completions route — extracted to app/routes/chat.py
+    # Routes — extracted to app/routes/*.py
     try:
         from app.routes.chat import router as chat_router, register_chat_helpers
         from app.routes.settings import router as settings_router, register_settings_helpers
+        from app.routes.status import router as status_router, register_status_helpers
+        from app.routes.runtime import router as runtime_router, register_runtime_helpers
+        from app.routes.models import router as models_router, register_models_helpers
     except ModuleNotFoundError:
         from routes.chat import router as chat_router, register_chat_helpers  # type: ignore[no-redef]
         from routes.settings import router as settings_router, register_settings_helpers  # type: ignore[no-redef]
+        from routes.status import router as status_router, register_status_helpers  # type: ignore[no-redef]
+        from routes.runtime import router as runtime_router, register_runtime_helpers  # type: ignore[no-redef]
+        from routes.models import router as models_router, register_models_helpers  # type: ignore[no-redef]
 
     register_chat_helpers(
         build_status=build_status,
@@ -2319,8 +1488,16 @@ def create_app(runtime: RuntimeConfig | None = None, enable_orchestrator: bool |
     register_settings_helpers(
         restart_managed_llama_process=restart_managed_llama_process,
     )
+    import sys
+    _this_module = sys.modules[__name__]
+    register_status_helpers(main_module=_this_module, chat_html_path=_CHAT_HTML_PATH)
+    register_runtime_helpers(main_module=_this_module)
+    register_models_helpers(main_module=_this_module)
     app.include_router(chat_router)
     app.include_router(settings_router)
+    app.include_router(status_router)
+    app.include_router(runtime_router)
+    app.include_router(models_router)
 
     return app
